Adding Glucose
Dropped in a statically linked Glucose binary we can call. This is a Linux executable, so you can run it there, or on the windows-linux-subsystem. Note that file path to is currently hard coded in the Java, so you'll probably need to change that if you want to run it.